backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
python-multipart==0.0.20
playwright==1.49.1
pytest==8.3.4
httpx==0.28.1
